100% accrat from floats to  cpu to gpu not close but perffect
* gpu to cpu is perrect match on gloats

* updated CPU patch only for cpu to cpu or anything which just keeps this

+++ CAESAR_decompressor.py
@@ -406,6 +406,7 @@ model = CompressorMix(
 
 state_dict = remove_module_prefix(torch.load('./pretrained/caesar_v.pt', map_location=device))
 model.load_state_dict(state_dict)
+model = model.double()
 
 quantized_cdf, cdf_length, offset = model.entropy_model.prior._update(30)
 medians = model.entropy_model.prior.medians.detach()
@@ -420,7 +421,7 @@ model.eval()
 with torch.no_grad():
     print('device: ', device)
     model = model.to(device)
-    example_inputs=(torch.randn(10, 2, 64, 16, 16, device=device),)
+    example_inputs=(torch.randn(10, 2, 64, 16, 16, device=device).double(),)
     batch_dim = torch.export.Dim("batch", min=1, max=512)
     # [Optional] Specify the first dimension of the input x as dynamic.
     exported = torch.export.export(model, example_inputs, dynamic_shapes={"x": {0: batch_dim}})
